Make parser pass conformance tests
This commit adds support for testing the conformance of the parser to an
array of specification tests. These tests are intended to be available
in a `specs/` directory at the top level of the repository. The specs
are written as a series of input files (in TOML) and output files (in
JSON). The JSON files conform to the interface provided in the TOML
project's compliance testing repository.

The commit contains numerous fixes to the parser to address bugs that
were revealed during the conformance tests.